Add arena for constructing arbitrary spans
The source-rewriting capabilities of the parsers would be incredibly
limited if parsers couldn't construct tokens containing arbitrary new or
modified strings -- like Rust's proc macros. Therefore, we tie an
`Arena` (from `typed-arena`) of `String`s to the input source, and allow
parsers to add arbitrary tokens to the arena.

Cassie helped guide the design and fixing a bunch of *incredibly*
annoying lifetime errors along the way.

Co-authored-by: Cassie Jones <code@witchoflight.com>